Modifications in initializa function
Initially the program was not able to find the intervals of the roots of a polynomial where all the terms are positive. However, after setting a flag variable in the initalize function(and passign the same to the operation function), we are able to detect such a equation and use negative values (starting from i=-5) to find the closed interval.